+++ pcs/app.py
@@ -11,7 +11,9 @@ import logging
 
 from pcs import (
     acl,
+    alert,
     booth,
+    client,
     cluster,
     config,
     constraint,
@@ -26,7 +28,6 @@ from pcs import (
     stonith,
     usage,
     utils,
-    alert,
 )
 
 from pcs.cli.common import (
@@ -202,6 +203,11 @@ def main(argv=None):
             argv,
             utils.get_modifiers()
         ),
+        "client": lambda argv: client.client_cmd(
+            utils.get_library_wrapper(),
+            argv,
+            utils.get_modifiers()
+        ),
     }
     if command not in cmd_map:
         usage.main()
@@ -254,7 +260,7 @@ def main(argv=None):
 
             # call the local pcsd
             err_msgs, exitcode, std_out, std_err = utils.call_local_pcsd(
-                orig_argv, True
+                orig_argv
             )
             if err_msgs:
                 for msg in err_msgs: